Develop controller functions
